move schema model into seperate folder
